file renameing as suggested by @f-allian
