+++ causal_testing/estimation/abstract_estimator.py
