+++ tests/estimation_tests/test_instrumental_variable_estimator.py
@@ -6,7 +6,7 @@ from causal_testing.specification.variable import Input
 from causal_testing.utils.validation import CausalValidator
 from causal_testing.specification.capabilities import TreatmentSequence
 
-from causal_testing.estimation.iv_estimator import InstrumentalVariableEstimator
+from causal_testing.estimation.instrumental_variable_estimator import InstrumentalVariableEstimator
 
 
 class TestInstrumentalVariableEstimator(unittest.TestCase):
