+++ tests/estimation_tests/test_linear_regression_estimator.py
@@ -7,7 +7,7 @@ from causal_testing.utils.validation import CausalValidator
 from causal_testing.specification.capabilities import TreatmentSequence
 
 from causal_testing.estimation.linear_regression_estimator import LinearRegressionEstimator
-from causal_testing.estimation.gp import reciprocal
+from causal_testing.estimation.genetic_programming_regression_fitter import reciprocal
 
 
 def load_nhefs_df():
